refactor: Position 디렉토리 변경
domain > domain/gameElements/Position
